rename, refactor, remove two-side relationship in model
