+++ src/main/java/com/artisan/solutions/gallery/persistence/repository/ArtistRepository.java
@@ -1,10 +1,10 @@
-package com.artisan.solutions.gallery.persistence.repo;
+package com.artisan.solutions.gallery.persistence.repository;
 
 import com.artisan.solutions.gallery.persistence.model.Artist;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArtistRepo extends JpaRepository<Artist, Long> {
+public interface ArtistRepository extends JpaRepository<Artist, Long> {
 
 }
