+++ src/main/java/com/artisan/solutions/gallery/persistence/repository/WorkRepository.java
@@ -1,10 +1,10 @@
-package com.artisan.solutions.gallery.persistence.repo;
+package com.artisan.solutions.gallery.persistence.repository;
 
 import com.artisan.solutions.gallery.persistence.model.Work;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkRepo extends JpaRepository<Work, Long> {
+public interface WorkRepository extends JpaRepository<Work, Long> {
 
 }
