+++ src/main/java/com/artisan/solutions/gallery/service/impl/ArtistServiceImpl.java
@@ -1,7 +1,7 @@
 package com.artisan.solutions.gallery.service.impl;
 
 import com.artisan.solutions.gallery.persistence.model.Artist;
-import com.artisan.solutions.gallery.persistence.repo.ArtistRepo;
+import com.artisan.solutions.gallery.persistence.repository.ArtistRepository;
 import com.artisan.solutions.gallery.service.ArtistService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,26 +12,26 @@ import java.util.Optional;
 @Service
 public class ArtistServiceImpl implements ArtistService {
 
-    private ArtistRepo artistRepo;
+    private ArtistRepository artistRepository;
 
     @Autowired
-    public ArtistServiceImpl(ArtistRepo artistRepo) {
-        this.artistRepo = artistRepo;
+    public ArtistServiceImpl(ArtistRepository artistRepository) {
+        this.artistRepository = artistRepository;
     }
 
     @Override
     public Artist create(Artist artist) {
-        return artistRepo.save(artist);
+        return artistRepository.save(artist);
     }
 
     @Override
     public Optional<Artist> findById(Long id) {
-        return artistRepo.findById(id);
+        return artistRepository.findById(id);
     }
 
     @Override
     public List<Artist> findAll() {
-        return artistRepo.findAll();
+        return artistRepository.findAll();
     }
 
     @Override
@@ -41,6 +41,6 @@ public class ArtistServiceImpl implements ArtistService {
 
     @Override
     public void deleteById(Long id) {
-        artistRepo.deleteById(id);
+        artistRepository.deleteById(id);
     }
 }
