+++ src/main/java/com/artisan/solutions/gallery/service/impl/WorkServiceImpl.java
@@ -1,7 +1,7 @@
 package com.artisan.solutions.gallery.service.impl;
 
 import com.artisan.solutions.gallery.persistence.model.Work;
-import com.artisan.solutions.gallery.persistence.repo.WorkRepo;
+import com.artisan.solutions.gallery.persistence.repository.WorkRepository;
 import com.artisan.solutions.gallery.service.WorkService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,35 +12,35 @@ import java.util.Optional;
 @Service
 public class WorkServiceImpl implements WorkService {
 
-    private WorkRepo workRepo;
+    private WorkRepository workRepository;
 
     @Autowired
-    public WorkServiceImpl(WorkRepo workRepo) {
-        this.workRepo = workRepo;
+    public WorkServiceImpl(WorkRepository workRepository) {
+        this.workRepository = workRepository;
     }
 
     @Override
     public Work create(Work work) {
-        return workRepo.save(work);
+        return workRepository.save(work);
     }
 
     @Override
     public Optional<Work> findById(Long id) {
-        return workRepo.findById(id);
+        return workRepository.findById(id);
     }
 
     @Override
     public List<Work> findAll() {
-        return workRepo.findAll();
+        return workRepository.findAll();
     }
 
     @Override
     public Work update(Work work) {
-        return workRepo.save(work);
+        return workRepository.save(work);
     }
 
     @Override
     public void deleteById(Long id) {
-        workRepo.deleteById(id);
+        workRepository.deleteById(id);
     }
 }
